🎨 NotificationSender, NotificationService 인터페이스로 분리
스크럼 때 논의한 대로 인터페이스로 분리
